Revamped home page, made org not found page

+++ app.py
@@ -14,7 +14,7 @@ app.config["SECRET_KEY"] = "2034tupgfbdsJI0--tqwhudbP9q5w[yyjokuui[lj[];m]/zxcb/
 @app.route("/", methods = ['GET','POST'])
 def org_page():
     db, cur = get_db()
-    warning = ""
+    message = ""
     show_confirmation = False
     if (request.method == 'POST'):
         neworgname = request.form.get("neworgname")
@@ -24,8 +24,11 @@ def org_page():
         #CASE: confirming deletion of organization. Organization name is stored in session["remove"]
         if removal_confirmation is not None:
             assert session.get("remove") is not None
-            cur.execute(f"DELETE FROM organization WHERE name = \"{session.get("remove")}\"")
-            db.commit()
+            if removal_confirmation == "Confirm":
+                cur.execute(f"DELETE FROM organization WHERE name = \"{session.get("remove")}\"")
+                db.commit()
+            else:
+                message = f"Removal of \"{session.get("remove")}\" has been cancelled."
 
         #CASE: creating new organization. 
         elif remove_org_name is None:
@@ -38,11 +41,16 @@ def org_page():
             
             #Invalid: Whitespace names
             if (len(neworgname) == 0):
-                warning = "Please enter a non-whitespace name."
+                message = "Please enter a non-whitespace name."
+                request.form.neworgname = ""
+
+            #Invalid: name is too long
+            elif (len(neworgname) > 255):
+                message = "Please enter an organization name shorter than 255 characters."
 
             #Invalid: org name is already used
             elif (len(org_name_overlap) != 0):
-                warning = "This organization name already exists; Please use a different name."
+                message = "This organization name already exists; Please use a different name."
             
             #Valid: New organization added, name field is cleared
             else:
@@ -53,28 +61,32 @@ def org_page():
 
         #CASE: requesting deletion of an organization. 
         else:
-            warning = f"Do you wish to remove \"{remove_org_name}\"? All associated contacts and contact info will be deleted."
+            message = f"Do you wish to remove \"{remove_org_name}\"? All associated contacts and contact info will be deleted."
             session["remove"] = remove_org_name
             show_confirmation = True
         
         
     
-    get_orgs = (cur.execute("SELECT name FROM organization")).fetchall()
+    get_orgs = (cur.execute("SELECT name FROM organization ORDER BY name")).fetchall()
 
-    return render_template("index.html", orgs = get_orgs, warn = warning, show = show_confirmation)
-
-
+    return render_template("index.html", orgs = get_orgs, warn = message, show = show_confirmation)
 
 
-@app.route("/<orgname>", methods = ['GET','POST'])
+
+
+@app.route("/organization/<orgname>", methods = ['GET','POST'])
 def contact_page(orgname):
     session.clear()
+    db, cur = get_db()
+    find_organization = (cur.execute(f"SELECT id FROM organization WHERE name = \"{orgname}\"")).fetchone()
+    if (find_organization is None):
+        return render_template("org_notfound.html",org_name = orgname)
     #see if this is a valid orgname
     #if it isn't, return an error page
     #We're sorry
     #the organization you're looking cannot be found. 
     #link returns to homepage
-    pass
+    return "Hello World"
 
 @app.teardown_appcontext
 def close_connection(exception):
